chore: Refactor in add alarm detail styles
